Update GC from CoreRT
https://github.com/dotnet/corert/tree/master/src/Native/gc dotnet/coreclr@102af78f9169082478165e7c676bbee6ae194940


Commit migrated from https://github.com/dotnet/coreclr/commit/71ccdf5472c97326305ee116198218e3c9df4c53

+++ src/clr/src/gc/env/gcenv.os.h
@@ -181,7 +181,7 @@ public:
     // current platform. It is indended for logging purposes only.
     // Return:
     //  Numeric id of the current thread or 0 if the 
-    static uint32_t GetCurrentThreadIdForLogging();
+    static uint64_t GetCurrentThreadIdForLogging();
 
     // Get id of the current process
     // Return:
